Support Windows with noimage mode (#112)
# Overview

This PR adds support to build AOTriton dispatcher library under Windows.

## Known Limitations

On Windows, `dladdr` is implemented through
[dlfcn-win32](https://github.com/dlfcn-win32/dlfcn-win32), which unfortunately
uses `GetModuleFileNameA` (ANSI version) to get the base directory of
`libaotriton.so`. This means, if `libaotriton.so` is put under a path with
characters that cannot represented in current code page, the loading of GPU
kernels will fail.

### Possible Workaround

* Globally set Code Page to 65001 (UTF-8)
* Specify [activeCodePage with Application
  manifests](https://learn.microsoft.com/en-us/windows/win32/sbscs/application-manifests#activeCodePage)
  to force the process using UTF-8

## Major Changes

* [build] Add Win32 support
* [doc] Update dependencies on Win32
* [shim] Introduce `pstring_view` to replace `std::string_view` as
  platform-dependent path string
* [gen] Generate UTF-16 string literals on Windows instead.
* [gen] Fix `KeyError: 'validator_function'` when none of the target GPUs
  support AITER ASM kernel.
* [shim] Add `packed_kernel_win32.h` and `packed_kernel_unix.h`
  
## Minor Changes
  
* [shim] Cast data types explicitly to suppress warnings.

+++ v3src/flash/attn_bwd.cc
@@ -127,7 +127,7 @@ attn_bwd(const attn_bwd_params& in,
     .CAUSAL_TYPE = in.causal_type,
     .ENABLE_DROPOUT = in.dropout_p > 0.0,
     .PADDED_HEAD = head_dim != head_dim_rounded,
-    .BIAS_TYPE = bool(in.B) ? 1 : 0,
+    .BIAS_TYPE = static_cast<int8_t>(bool(in.B) ? 1 : 0),
   };
   OpAttnBwdContext context;
   context.params = &params;
